+++ src/net/gameserver/player.cpp
@@ -39,90 +39,90 @@ void Net::GameServer::Player::say(const std::string &text)
 void Net::GameServer::Player::walk(int x, int y)
 {
     MessageOut msg(PGMSG_WALK);
-    msg.writeShort(x);
-    msg.writeShort(y);
+    msg.writeInt16(x);
+    msg.writeInt16(y);
     Net::GameServer::connection->send(msg);
 }
 
 void Net::GameServer::Player::pickUp(int x, int y)
 {
     MessageOut msg(PGMSG_PICKUP);
-    msg.writeShort(x);
-    msg.writeShort(y);
+    msg.writeInt16(x);
+    msg.writeInt16(y);
     Net::GameServer::connection->send(msg);
 }
 
 void Net::GameServer::Player::moveItem(int oldSlot, int newSlot, int amount)
 {
     MessageOut msg(PGMSG_MOVE_ITEM);
-    msg.writeByte(oldSlot);
-    msg.writeByte(newSlot);
-    msg.writeByte(amount);
+    msg.writeInt8(oldSlot);
+    msg.writeInt8(newSlot);
+    msg.writeInt8(amount);
     Net::GameServer::connection->send(msg);
 }
 
 void Net::GameServer::Player::drop(int slot, int amount)
 {
     MessageOut msg(PGMSG_DROP);
-    msg.writeByte(slot);
-    msg.writeByte(amount);
+    msg.writeInt8(slot);
+    msg.writeInt8(amount);
     Net::GameServer::connection->send(msg);
 }
 
 void Net::GameServer::Player::equip(int slot)
 {
     MessageOut msg(PGMSG_EQUIP);
-    msg.writeByte(slot);
+    msg.writeInt8(slot);
     Net::GameServer::connection->send(msg);
 }
 
 void Net::GameServer::Player::unequip(int slot)
 {
     MessageOut msg(PGMSG_UNEQUIP);
-    msg.writeByte(slot);
+    msg.writeInt8(slot);
     Net::GameServer::connection->send(msg);
 }
 
 void Net::GameServer::Player::useItem(int slot)
 {
     MessageOut msg(PGMSG_USE_ITEM);
-    msg.writeByte(slot);
+    msg.writeInt8(slot);
     Net::GameServer::connection->send(msg);
 }
 
 void Net::GameServer::Player::attack(int direction)
 {
     MessageOut msg(PGMSG_ATTACK);
-    msg.writeByte(direction);
+    msg.writeInt8(direction);
     Net::GameServer::connection->send(msg);
 }
 
 void Net::GameServer::Player::changeAction(Being::Action action)
 {
     MessageOut msg(PGMSG_ACTION_CHANGE);
-    msg.writeByte(action);
+    msg.writeInt8(action);
     Net::GameServer::connection->send(msg);
 }
 
 void Net::GameServer::Player::talkToNPC(int id, bool restart)
 {
     MessageOut msg(restart ? PGMSG_NPC_TALK : PGMSG_NPC_TALK_NEXT);
-    msg.writeShort(id);
+    msg.writeInt16(id);
     Net::GameServer::connection->send(msg);
 }
 
 void Net::GameServer::Player::selectFromNPC(int id, int choice)
 {
     MessageOut msg(PGMSG_NPC_SELECT);
-    msg.writeShort(id);
-    msg.writeByte(choice);
+    msg.writeInt16(id);
+    msg.writeInt8(choice);
     Net::GameServer::connection->send(msg);
 }
 
 void Net::GameServer::Player::requestTrade(int id)
 {
     MessageOut msg(PGMSG_TRADE_REQUEST);
-    msg.writeShort(id);
+    msg.writeInt16(id);
     Net::GameServer::connection->send(msg);
 }
 
@@ -135,22 +135,22 @@ void Net::GameServer::Player::acceptTrade(bool accept)
 void Net::GameServer::Player::tradeItem(int slot, int amount)
 {
     MessageOut msg(PGMSG_TRADE_ADD_ITEM);
-    msg.writeByte(slot);
-    msg.writeByte(amount);
+    msg.writeInt8(slot);
+    msg.writeInt8(amount);
     Net::GameServer::connection->send(msg);
 }
 
 void Net::GameServer::Player::tradeMoney(int amount)
 {
     MessageOut msg(PGMSG_TRADE_SET_MONEY);
-    msg.writeLong(amount);
+    msg.writeInt32(amount);
     Net::GameServer::connection->send(msg);
 }
 
 void Net::GameServer::Player::tradeWithNPC(int item, int amount)
 {
     MessageOut msg(PGMSG_NPC_BUYSELL);
-    msg.writeShort(item);
-    msg.writeShort(amount);
+    msg.writeInt16(item);
+    msg.writeInt16(amount);
     Net::GameServer::connection->send(msg);
 }
